be more tolerant of extra log messages
When waiting for the server to initialize it tails the log file starting at the last line, if two log messages come in at the same time, then it may not see the "initialized" and hang indefinitely. Sometimes a GC message appears and it see's that as the last line. Start tailing the log at the last couple of lines instead of the first line and start tailing the log sooner